Portability: fix cxxopts on GCC15, MSVC std::max mismatch, Windows setenv/realpath
Three more gaps surfaced now that CI finally compiles regenie's own sources
on Ubuntu and Windows for the first time:

cxxopts.hpp (Ubuntu, GCC 15): uses uint8_t/intmax_t without including
<cstdint> itself, relying on transitive inclusion that GCC 15's libstdc++
no longer provides on this path. Added the include explicitly before
cxxopts.hpp, the only place it's included.

check_sets_include_exclude (Windows, MSVC): bsize was declared 'unsigned
long', which is 32-bit under Windows' LLP64 model, then compared via
std::max against a vector's size_type (size_t, 64-bit even on Windows).
MSVC's std::max requires an exact type match; GCC/Clang's libstdc++/libc++
happened to accept it because unsigned long and size_t are the same width
under Linux/macOS's LP64 model, masking the bug there. Changed to size_t,
which is what it's actually compared against; identical width and behavior
on Linux/macOS, so no change there.

get_fullpath (Windows): unconditionally called setenv() and realpath(),
neither of which exist on Windows. The locale bug setenv(LC_ALL, C) works
around is glibc/Boost-specific and doesn't occur on Windows, so that whole
retry branch is skipped there; realpath() is replaced with _fullpath(),
Windows' equivalent, behind #ifdef _WIN32 -- POSIX behavior for
Linux/macOS is untouched.

Verified: local build unaffected (identical step 1 LOCO output to the
pre-change baseline); grepped the rest of src/ for other POSIX-only calls
(setenv/realpath/mkdir/fork/popen/pthread_*/sysconf/getrusage/unlink) and
found no other unguarded uses.

+++ src/Geno.cpp
@@ -3124,7 +3124,7 @@ void read_setlist(const struct in_files* files, struct param* params, struct fil
 void check_sets_include_exclude(bool const& bsize_set, const struct in_files* files, struct param* params, struct filter* filters, vector< vector<vset> >& setinfo, mstream& sout){
 
   uint32_t nsets = 0;
-  unsigned long bsize = 0;
+  size_t bsize = 0;
   vector< vector<vset> > tmp_setinfo;
   map<string, vector<int> >::iterator itr;
